Debugged setup.py
There were 2 bugs:

- The version number was using the module version number, which
required to import the module. When importing the module for the first
time (during install), the config file would not exist and the setup
would crash. I replaced it with a constant version number.

- On my raspberry pi, when using sudo the user name and home directory
were not correct (/root/ instead of /home/pi/). The config file would
then be saved in the /root directory and not in the user directory. I
used $SUDO_USER to figure out the real user name and os.path.expander
to get the user home directory.